Better parallelization w.r.t. incoming results from other processes.
mapsolvers are now "synchronized" (a la Java) to allow the results
receiving thread to add blocking clauses as it gets results, instead of
having to store them indefinitely waiting for MUSer to finish.  Several
changes were required to make this work correctly, including the removal
of any direct access to an msolver's internals by other classes.  Only
msolvers' *methods* are protected by a lock, so that must be the only
way msolvers are used in parallel code.